+++ src/client/components/JoinOrganizationModal.tsx
@@ -1,13 +1,16 @@
 /** @jsx jsx */
 import AriaModal from "react-aria-modal";
 import { connect } from "react-redux";
+
 import { Box, jsx, ThemeUIStyleObject } from "theme-ui";
 
-import { IOrganization } from "../../shared/entities";
+import { IOrganization, IUser } from "../../shared/entities";
 import { showCopyMapModal } from "../actions/districtDrawing";
 import { State } from "../reducers";
 import store from "../store";
 import { AuthModalContent } from "./AuthComponents";
+import ConfirmJoinOrganization from "./ConfirmJoinOrganization";
+import { Resource } from "../resource";
 
 const style: ThemeUIStyleObject = {
   footer: {
@@ -34,10 +37,12 @@ const style: ThemeUIStyleObject = {
 
 const JoinOrganizationModal = ({
   organization,
+  user,
   showModal
 }: {
   readonly organization: IOrganization;
   readonly showModal: boolean;
+  readonly user: Resource<IUser>;
 }) => {
   const hideModal = () => store.dispatch(showCopyMapModal(false));
 
@@ -49,7 +54,13 @@ const JoinOrganizationModal = ({
       getApplicationNode={() => document.getElementById("root") as Element}
       underlayStyle={{ paddingTop: "4.5rem" }}
     >
-      <Box sx={style.modal}>{<AuthModalContent organization={organization} />}</Box>
+      <Box sx={style.modal}>
+        {"resource" in user ? (
+          <ConfirmJoinOrganization organization={organization} />
+        ) : (
+          <AuthModalContent organization={organization} />
+        )}
+      </Box>
     </AriaModal>
   ) : null;
 };